Split Hi section into separate component
