+++ src/App.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import Header from "./components/Header.vue";
 import Footer from "./components/Footer.vue";
-import Content from "./components/Content.vue";
+import Content from "./components/Content/Content.vue";
 </script>
 
 <template>
